Separate ModMenu entry and ScreenBuilder, so we can display config even if no ModMenu available
